Allow sampling of the max fuzz value
The end value of the range function is exclusive. In order to allow
sampling of the value corresponding fo fuzz_max, we need to add 1 here.
E.g., the default values of

  fuzz_min: 0
  fuzz_max: 4

should result in the following list

  [0, 1, 2, ..., 398, 399, 400]

This has the added benefit of also allowing fuzz_min = fuzz_max, to add
an deterministic border around the objects.