Always use fuzzing when specified
There are a few cases to consider when thinking about fuzzing

1) No objects loaded
2) Full plate of objects loaded
3) Loaded objects are close to the mesh limits
4) Loaded objects well within the mesh limits

The current implementation does no fuzzing in cases 1 and 2. Does
limited fuzzing (between object limits and mesh limits) in case 3. And
only in case 4 is the fuzzing performed over an interval of size
fuzz_max - fuzz_min.

The new implementation does fuzzing over an interval of fuzz_max -
fuzz_min in all 4 cases.

Looking at x_min, the new method fuzzes between
  [bed_mesh_min[0], bed_mesh_min[0] + fuzz_max - fuzz_min]
for case 1, 2 and 3. And between
  [minimum_x - fuzz_max, minimum_x - fuzz_min]
for case 4 (I.e., when the minimum x point is at least fuzz_max from the
mesh limit).

This fullfills the following goals:
- Never probe outsize of bed_mesh min / max.
- Always fuzz over an interval of size fuzz_max - fuzz_min.

The above logic change additionally allows to remove the now redundant
condition (if all_points) and to extract the common min / max
computation between fuzzing and not fuzzing.